+++ waspc/e2e-tests/Main.hs
@@ -21,6 +21,7 @@ import Tests.WaspCompletionTest (waspCompletionTest)
 import Tests.WaspDbMigrateDevTest (waspDbMigrateDevTest)
 import Tests.WaspDbResetTest (waspDbResetTest)
 import Tests.WaspDbSeedTest (waspDbSeedTest)
+import Tests.WaspDbStartTest (waspDbStartTest)
 import Tests.WaspDepsTest (waspDepsTest)
 import Tests.WaspDockerfileTest (waspDockerfileTest)
 import Tests.WaspInstallTest (waspInstallTest)
@@ -113,7 +114,7 @@ e2eTests = do
         waspDockerfileTest,
         -- FIXME: waspStudioTest,
         -- Wasp project db commands
-        -- FIXME: waspDbStartTest,
+        waspDbStartTest,
         -- FIXME: waspDbStudioTest,
         waspDbSeedTest,
         waspDbResetTest,