+++ waspc/src/Wasp/Db/Postgres.hs
@@ -1,23 +1,28 @@
 module Wasp.Db.Postgres
   ( makeConnectionUrl,
     postgresMaxDbNameLength,
+    defaultPostgresPort,
     defaultPostgresDockerImageSpec,
   )
 where
 
+import Network.Socket (PortNumber)
 import Text.Printf (printf)
 import Wasp.Util.Docker
   ( DockerImageName,
     DockerVolumeMountPath,
   )
 
-makeConnectionUrl :: String -> String -> Int -> String -> String
+makeConnectionUrl :: String -> String -> PortNumber -> String -> String
 makeConnectionUrl user pass port dbName =
-  printf "postgresql://%s:%s@localhost:%d/%s" user pass port dbName
+  printf "postgresql://%s:%s@localhost:%s/%s" user pass (show port) dbName
 
 -- As specified by PostgreSQL documentation.
 postgresMaxDbNameLength :: Int
 postgresMaxDbNameLength = 63
+
+defaultPostgresPort :: PortNumber
+defaultPostgresPort = 5432
 
 -- | We pin the Postgres Docker image to avoid issues when a new major version of Postgres
 -- is released. We aim to occasionally update this version in Wasp releases.